% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-cosine-signal.R
\name{c2kp}
\alias{c2kp}
\title{Basis C_{k, 1, p} in fitting procedure}
\usage{
c2kp(p, n, N, f, k, S, L, i)
}
\arguments{
\item{p}{window position}

\item{n}{window size}

\item{N}{lenth of original time-series}

\item{f}{frequency: number of cycles in length n window}

\item{k}{fourier frequency we are searching}

\item{S}{start of local signal}

\item{L}{length of local signal}

\item{i}{imaginary number}
}
\description{
Basis C_{k, 1, p} in fitting procedure
}
