% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-cosine-signal.R
\name{coef_to_search}
\alias{coef_to_search}
\title{Choose frequency time-series in SWDFT to search}
\usage{
coef_to_search(b, n)
}
\arguments{
\item{b}{2D complex array from the 'swdft' function}

\item{n}{window size}
}
\value{
k: the frequency time-series to search
}
\description{
Choose frequency time-series in SWDFT to search
}
