% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine-regression.R
\name{coefficients.swdft_mod}
\alias{coefficients.swdft_mod}
\title{Coefficients method for swdft_cosreg objects}
\usage{
\method{coefficients}{swdft_mod}(x, ...)
}
\arguments{
\item{x}{A swdft_cosreg object}
}
\description{
Coefficients method for swdft_cosreg objects
}
