% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-cosine-signal.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute MSE of local fit given frequency F and grid-point}
\usage{
compute_mse(f, S, L, bk, n, N, p_range, k, i, return_amp_phase = FALSE)
}
\arguments{
\item{f}{frequency: number of cycles in length n window}

\item{S}{start of local signal}

\item{L}{length of local signal}

\item{bk}{frequency time-series to search}

\item{n}{window size}

\item{N}{length of original time-series}

\item{p_range}{range of window positions}

\item{k}{fourier frequency we are searching}

\item{i}{imaginary number}

\item{return_amp_phase}{should we return  *just* the mse, or the mse plus
the parameters estimates. Used because we need a single number output
for optimization functions.}
}
\description{
Compute MSE of local fit given frequency F and grid-point
}
