% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{cosine}
\alias{cosine}
\title{Cosine signal with adjustable parameters}
\usage{
cosine(N, A = 1, Fr = 1, phase = 0)
}
\arguments{
\item{N}{signal length}

\item{A}{Amplitude}

\item{Fr}{Frequency: Number of cycles in a length N period}

\item{phase}{phase}
}
\value{
numeric vector with cosine function of x
}
\description{
Cosine signal with adjustable parameters
}
