% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapers.R
\name{cosine_taper}
\alias{cosine_taper}
\title{Cosine bell data taper}
\usage{
cosine_taper(n, p = 0.1)
}
\arguments{
\item{n}{length of time-series to taper}

\item{p}{proportion of ends to taper}
}
\value{
length n cosine bell taper w/ proportion p
}
\description{
Cosine bell data taper
}
