% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine-regression.R
\name{cosreg}
\alias{cosreg}
\title{Cosine regression}
\usage{
cosreg(x, f)
}
\arguments{
\item{x}{numeric. Signal.}

\item{f}{numeric. scalar or vector of frequencies to fit.}
}
\value{
S3 object of class 'swdft_cosreg'. See ?new_swdft_cosreg for details.
}
\description{
Cosine regression
}
