% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_swdft_cnum}
\alias{cov_swdft_cnum}
\title{Covariance between two complex-numbered outputs}
\usage{
cov_swdft_cnum(k, l, delta, n, sigma)
}
\arguments{
\item{k}{frequency of first coefficient}

\item{l}{frequency of second coefficient}

\item{delta}{window position shift of second coefficient}

\item{n}{window size}

\item{sigma}{white noise standard error}
}
\value{
complex-valued number of the covariance
}
\description{
Covariance between two complex-numbered outputs
}
