% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demod.R
\name{demod_swdft}
\alias{demod_swdft}
\title{Demodulate a Fourier Frequency with the SWDFT}
\usage{
demod_swdft(a, k)
}
\arguments{
\item{a}{swdft}

\item{k}{frequency to demodulate}
}
\description{
Demodulate a Fourier Frequency with the SWDFT
}
