% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet}
\alias{dirichlet}
\title{Dirichlet Kernel (Weight) for arbitrary summation indices}
\usage{
dirichlet(x, phase = 0, a = 0, b = length(x) - 1)
}
\arguments{
\item{x}{numeric to evaluate}

\item{phase}{defaults to 0}

\item{a}{start of summation index}

\item{b}{end of summation index}
}
\value{
sum of a complex exponential sum
}
\description{
Dirichlet Kernel (Weight) for arbitrary summation indices
}
