% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet_kernel}
\alias{dirichlet_kernel}
\title{Dirichlet Kernel}
\usage{
dirichlet_kernel(x, n, dw = FALSE)
}
\arguments{
\item{x}{variable evaluated by dirichlet kernel}

\item{n}{size of Dirichlet kernel}

\item{dw}{logical whether to add the Dirichlet Weight (DW) factor}
}
\value{
evaluation of the Dirichlet Kernel (D_n(x))
}
\description{
Dirichlet Kernel
}
