% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{get_aphi}
\alias{get_aphi}
\title{Extract amplitude and phase}
\usage{
get_aphi(x, S, L, f)
}
\arguments{
\item{x}{signal}

\item{S}{start parameter}

\item{L}{length pe}

\item{f}{frequency}
}
\description{
Extract amplitude and phase
}
