% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{get_freq_range}
\alias{get_freq_range}
\title{Get range of frequencies to search}
\usage{
get_freq_range(a, kwidth)
}
\arguments{
\item{a}{2D complex-valued array. The SWDFT to search}

\item{kwidth}{integer. the width of frequencies to search}
}
\description{
Get range of frequencies to search
}
