% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-cosine-signal.R
\name{get_grid}
\alias{get_grid}
\title{Compute possible start/length values of local signal}
\usage{
get_grid(N, n, lmin)
}
\arguments{
\item{N}{Length of original signal}

\item{n}{Window Size of SWDFT}

\item{lmin}{Minimum signal to search for}
}
\description{
Compute possible start/length values of local signal
}
