% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{get_loglik}
\alias{get_loglik}
\title{Compute the log likelihood}
\usage{
get_loglik(x, fitted, sigma, N)
}
\arguments{
\item{x}{signal}

\item{fitted}{fitted values}

\item{sigma}{estimated standard deviation}

\item{N}{length of x}
}
\description{
Compute the log likelihood
}
