% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_max_freq}
\alias{get_max_freq}
\title{Get the maximum DFT coefficient}
\usage{
get_max_freq(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric of largest frequency. Will be between 0 and .5
}
\description{
Get the maximum DFT coefficient
}
