% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{get_p_range}
\alias{get_p_range}
\title{Get range of P's to search}
\usage{
get_p_range(phat, n, N, pwidth, type = "around_max")
}
\arguments{
\item{phat}{integer. Window position with largest SWDFT coefficient}

\item{n}{integer. window size}

\item{N}{integer. Signal length}

\item{pwidth}{integer. the range of window positions to search for each window size}

\item{type}{character. either 'around max' or 'fullp'.}
}
\description{
Get range of P's to search
}
