% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{get_sigma}
\alias{get_sigma}
\title{Extract estimator of sigma}
\usage{
get_sigma(x, fitted, N)
}
\arguments{
\item{x}{signal}

\item{fitted}{fitted values}

\item{N}{length of x}
}
\description{
Extract estimator of sigma
}
