% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapers.R
\name{get_taper}
\alias{get_taper}
\title{Create taper for the SWDFT}
\usage{
get_taper(n, taper, p)
}
\arguments{
\item{n}{window size}

\item{taper}{taper type. Can be either 'none' (default) or 'cosine'}

\item{p}{proportion to taper on each end, if cosine taper is used}
}
\value{
length n taper
}
\description{
Create taper for the SWDFT
}
