% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-cosine-signal.R
\name{grid_search}
\alias{grid_search}
\title{Estimate Parameters for all Start/Length Combinations of Local Signal}
\usage{
grid_search(SL, bk, N, n, p_range, k, i)
}
\arguments{
\item{SL}{Vector of S and L parameters}

\item{bk}{frequency time-series to search}

\item{N}{length of original time-series}

\item{n}{window size}

\item{p_range}{range of window positions}

\item{k}{fourier frequency we are searching}

\item{i}{imaginary number}
}
\description{
Estimate Parameters for all Start/Length Combinations of Local Signal
}
