% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{lcr_loglik}
\alias{lcr_loglik}
\title{Log Likelihood}
\usage{
lcr_loglik(f, x, S, L, ftype = "full")
}
\arguments{
\item{f}{frequency}

\item{x}{signal}

\item{S}{start parameter}

\item{L}{length pe}

\item{ftype}{what to return}
}
\description{
Log Likelihood
}
