% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{local_cosreg}
\alias{local_cosreg}
\title{Local cosine regression}
\usage{
local_cosreg(x, lmin = 6, pwidth = 5, kwidth = 1, verbose = FALSE)
}
\arguments{
\item{x}{numeric signal to apply local cosine regression on}

\item{lmin}{integer. minimum signal length (L parameter) to search}

\item{pwidth}{integer. the range of window positions to search for each window size}

\item{kwidth}{integer. the width of frequencies to search}

\item{verbose}{logical. whether or not to print intermediate results}
}
\value{
S3 object of class 'swdft_local_cosreg'
}
\description{
Local cosine regression
}
