% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{local_signal}
\alias{local_signal}
\title{Local Periodic Signal}
\usage{
local_signal(N, A = 1, Fr = 1, phase = 0, S = 0, L = N)
}
\arguments{
\item{N}{signal length}

\item{A}{Amplitude}

\item{Fr}{Frequency: Number of cycles in a length N period}

\item{phase}{phase}

\item{S}{start of local signal}

\item{L}{length of local signal}
}
\value{
length N local periodic signal
}
\description{
Local Periodic Signal
}
