% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{moving_average}
\alias{moving_average}
\title{Simple high pass filter}
\usage{
moving_average(x, order)
}
\arguments{
\item{x}{the vector or time-series}

\item{order}{the order of the filter}
}
\description{
Simple high pass filter
}
