% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft2d.R
\name{new_swdft2d}
\alias{new_swdft2d}
\title{Constructor function for class 'swdft2d'}
\usage{
new_swdft2d(a, x, n0, n1, type)
}
\arguments{
\item{a}{4D complex-valued array of 2D SWDFT coefficients}

\item{x}{2D real or complex valued signal}

\item{n0}{window size in row direction}

\item{n1}{window size in column direction}

\item{type}{algorithm to implement. defaults to "fftw", other option 'fft' for R's base FFT function.
R's base fft function is used if}
}
\value{
S3 object w/ the same elements as arguments to this constructor function
}
\description{
Constructor function for class 'swdft2d'
}
