% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft3d.R
\name{new_swdft3d}
\alias{new_swdft3d}
\title{Constructor function for class 'swdft3d'}
\usage{
new_swdft3d(a, x, n0, n1, n2, type)
}
\arguments{
\item{a}{4D complex-valued array of 2D SWDFT coefficients}

\item{x}{3D real or complex-valued array}

\item{n0}{window size in dimension 0}

\item{n1}{window size in dimension 1}

\item{n2}{window size in dimension 2}

\item{type}{detaults to 'base', which is the only option}
}
\value{
S3 object w/ the same elements as arguments to this constructor function
}
\description{
Constructor function for class 'swdft3d'
}
