% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-cosine-regression.R
\name{new_swdft_local_cosreg}
\alias{new_swdft_local_cosreg}
\title{Constructor function for class 'swdft_local_cosreg'}
\usage{
new_swdft_local_cosreg(coefficients, fitted, residuals, data,
  window_params)
}
\arguments{
\item{coefficients}{matrix of coefficients for cosine regression model}

\item{fitted}{fitted values of cosine regression model}

\item{residuals}{residuals of cosine regression model}

\item{data}{original signal used to fit cosine regression}

\item{window_params}{data frame of fitted coefficients for each window size}
}
\value{
list with the following elements
\itemize{
  \item coefficients. A matrix of parameters, the three columns are: 1. amplitude 2. phase, and 3. frequency.
  There is only more that one row used when multiple frequencies are fit sequentially.
  \item fitted. fitted values of cosine regression model
  \item residuals. residuals of cosine regression model
  \item data. original signal used to fit cosine regression
  \item window_params. data frame of fitted coefficients for each window size
}
}
\description{
Constructor function for class 'swdft_local_cosreg'
}
