% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.swdft_mod}
\alias{plot.swdft_mod}
\title{Plot method for swdft_mod object}
\usage{
\method{plot}{swdft_mod}(x, y, ...)
}
\arguments{
\item{x}{A swdft_cosreg object}

\item{y}{not used, but required by plot generic function}
}
\description{
Plot method for swdft_mod object
}
