% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mvts}
\alias{plot_mvts}
\title{Multivariate Time-Series plot of the SWDFT}
\usage{
plot_mvts(a, take_log = FALSE, log_thresh = 0.01, only_unique = FALSE,
  legend = FALSE, title = "SWDFT Frequency Time-Series", cex_leg = 1,
  ylim = NULL, colors = NULL)
}
\arguments{
\item{a}{2D complex array of output of the 'swdft' function}

\item{take_log}{logical. whether to take the logarithm before plotting}

\item{log_thresh}{number threshhold for logs (so we don't get log(.00001) = -13)}

\item{only_unique}{logical. Only use the unique Fourier frequencies, not their aliases.}

\item{legend}{logical whether to include a legend}

\item{title}{Custom title}

\item{cex_leg}{size of the legend}

\item{ylim}{Custom y-limits}

\item{colors}{custom vector of colots}
}
\description{
Multivariate Time-Series plot of the SWDFT
}
\examples{
x <- rnorm(n = 100)
a <- swdft(x, n = 2^5)
plot_mvts(a)

}
