% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_swdft}
\alias{plot_swdft}
\title{Time-Frequency Plot of the SWDFT}
\usage{
plot_swdft(a, type = "Mod", use_fields = TRUE, zlim = NULL,
  xlab = "Window Position", ylab = "Frequency (Cycles/Window)",
  title = "SWDFT")
}
\arguments{
\item{a}{2D complex array of output of the 'swdft' function}

\item{type}{type of complex-number ourput. Either 'Re',
'Im', Arg', or 'Mod'. Defaults to 'Mod'}

\item{use_fields}{logical that determines whether we use image or
image.plot from the fields package. The key advantage of fields is
that it automatically provides a legend}
}
\description{
Time-Frequency Plot of the SWDFT
}
\examples{
x <- rnorm(n = 100)
a <- swdft(x, n = 2^5)
plot_swdft(a)

}
