% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine-regression.R
\name{residuals.swdft_mod}
\alias{residuals.swdft_mod}
\title{Residuals method for swdft_cosreg objects}
\usage{
\method{residuals}{swdft_mod}(object, ...)
}
\arguments{
\item{object}{A swdft_cosreg object}

\item{...}{optional arguments to match generic function}
}
\description{
Residuals method for swdft_cosreg objects
}
