% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{sine}
\alias{sine}
\title{Sine signal with adjustable parameters}
\usage{
sine(N, A = 1, Fr = 1, phase = 0)
}
\arguments{
\item{N}{length signal}

\item{A}{Amplitude}

\item{Fr}{Frequency: Number of cycles in a length N period}

\item{phase}{phase}
}
\value{
numeric vector with sine
}
\description{
Sine signal with adjustable parameters
}
