% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{smooth_pgram}
\alias{smooth_pgram}
\title{Smooth SWDFT coefficients with a convolution}
\usage{
smooth_pgram(a, fft_weight = NULL)
}
\arguments{
\item{a}{real-valued length n periodogram}

\item{fft_weight}{optionally specify the pre-computed FFT of the weights}
}
\value{
smoothed coefficients
}
\description{
Smooth SWDFT coefficients with a convolution
}
