% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{smooth_swdft}
\alias{smooth_swdft}
\title{Smooth the SWDFT coefficients}
\usage{
smooth_swdft(a, ktype = "daniell", m = 2, num_convs = 1)
}
\arguments{
\item{a}{real or complex-valued swdft. If real-valued, then we assume it's the squared
modules already. If it's complex valued, we convert to the squared modulus.}

\item{ktype}{either 'daniell' or 'modified.daniell'}

\item{m}{kernel width from stats::kernel}

\item{num_convs}{num_convs from stats::kernel}
}
\value{
Smooth squared modulues SWDFT coefficients
}
\description{
Smooth the SWDFT coefficients
}
