% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star}
\alias{star}
\title{Magnitude of a Variable Star}
\format{Length 600 numeric time-series}
\source{
\url{https://www.york.ac.uk/depts/maths/data/ts/welcome.htm}
}
\usage{
star
}
\description{
Magnitude of a Variable Star
}
\keyword{datasets}
