% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft2d.R
\name{swdft2d}
\alias{swdft2d}
\title{2D Sliding Window Discrete Fourier Transform}
\usage{
swdft2d(x, n0, n1, type = "fftw")
}
\arguments{
\item{x}{2D input signal}

\item{n0}{window size in row direction}

\item{n1}{window size in column direction}

\item{type}{algorithm to implement. defaults to "fftw", other option 'fft' for R's base FFT function.
R's base fft function is used if 'fftwtools' library is not installed.}
}
\value{
An S3 'swdft2d' object. See ?new_swdft for details.
}
\description{
2D Sliding Window Discrete Fourier Transform
}
