% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft2d.R
\name{swdft2d_fftw}
\alias{swdft2d_fftw}
\title{2D Sliding Window Discrete Fourier Transform using fftw}
\usage{
swdft2d_fftw(x, n0, n1)
}
\arguments{
\item{x}{2D input signal}

\item{n0}{window size in row direction}

\item{n1}{window size in column direction}
}
\description{
2D Sliding Window Discrete Fourier Transform using fftw
}
