% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft3d.R
\name{swdft3d}
\alias{swdft3d}
\title{3D Sliding Window Discrete Fourier Transform}
\usage{
swdft3d(x, n0, n1, n2, type = "base")
}
\arguments{
\item{x}{3D real or complex-valued array}

\item{n0}{window size in dimension 0}

\item{n1}{window size in dimension 1}

\item{n2}{window size in dimension 2}

\item{type}{defaults to 'base', which is the only option}
}
\value{
An S3 'swdft3d' object. See ?new_swdft for details.
}
\description{
3D Sliding Window Discrete Fourier Transform
}
