% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft3d.R
\name{swdft_base_3d}
\alias{swdft_base_3d}
\title{3D SWDFT using base R}
\usage{
swdft_base_3d(x, n0, n1, n2)
}
\arguments{
\item{x}{3D real or complex-valued array}

\item{n0}{window size in dimension 0}

\item{n1}{window size in dimension 1}

\item{n2}{window size in dimension 2}
}
\description{
3D SWDFT using base R
}
