% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft.R
\name{swdft_fftw}
\alias{swdft_fftw}
\title{Sliding Window Discrete Fourier Transform using fftw}
\usage{
swdft_fftw(x, n, taper)
}
\arguments{
\item{x}{real or complex vector}

\item{n}{integer window size.}

\item{taper}{type of taper for each window position. defaults to 'none', can also be 'cosine'.}
}
\value{
n x P array, where P = length(x) - n + 1
}
\description{
Sliding Window Discrete Fourier Transform using fftw
}
