% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swdft_to_props}
\alias{swdft_to_props}
\title{Convert the SWDFT to proportions of frequency}
\usage{
swdft_to_props(a)
}
\arguments{
\item{a}{swdft}
}
\description{
Convert the SWDFT to proportions of frequency
}
