% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine-regression.R
\name{fitted.swdft_mod}
\alias{fitted.swdft_mod}
\title{Fitted values method for swdft_cosreg objects}
\usage{
\method{fitted}{swdft_mod}(x, ...)
}
\arguments{
\item{x}{A swdft_cosreg object}
}
\description{
Fitted values method for swdft_cosreg objects
}
