% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdft.R
\name{swdft_fft}
\alias{swdft_fft}
\title{Sliding Window Discrete Fourier Transform with base R}
\usage{
swdft_fft(x, n)
}
\arguments{
\item{x}{real or complex vector}

\item{n}{window size}
}
\value{
n x P array, where P = length(x) - n + 1
}
\description{
Sliding Window Discrete Fourier Transform with base R
}
